refactor: move apis and queries to src root level
