+++ frontend/src/domains/Projects/components/CreateProjectModal/CreateProjectModal.tsx
@@ -5,7 +5,7 @@ import { Folder } from 'lucide-react';
 import {
   postCreateProjectMutationOptions,
   getProjectListQueryOptions,
-} from '@/shared/queries';
+} from '@/queries';
 import { Modal, Input, Button, AsyncBoundary } from '@/shared/components';
 import type { CreateProjectBody } from '@lgtmai/backend/types';
 import { FolderBrowser } from '../FolderBrowser/FolderBrowser';
