+++ frontend/src/queries/auth.ts
